project polished

+++ src/test/java/feature/TestClass.java
@@ -1,4 +1,5 @@
-import cucumber.api.java.en.Given;
+package feature;
+
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
 import org.junit.*;
@@ -29,10 +30,10 @@ public class TestClass {
     Opens the driver and goes to the url given.  Checks whether it is loaded.
     throws Assertion Error if not
      */
-    @When("^I have opened \"([^\"]*)\" as a browser$")
+
+    @When("^I have opened \"([^\"]*)\" in a browser$")
     public void setUp(String baseUrl) {
         driver = new FirefoxDriver();
-        baseUrl = new String("https://www.amazon.com/");
         driver.get(baseUrl);
         driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
 
@@ -47,8 +48,6 @@ public class TestClass {
 
     @When("^I sign in with ([^/]+)/([^/]+)$")
     public void testAmazonSignIn(String eMailAddress, String password) {
-        eMailAddress = new String("hazalgunduz@yahoo.com");
-        password = new String("123456");
 
         driver.findElement(By.cssSelector("#nav-link-accountList > span:nth-child(1)")).click();
         driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
@@ -63,24 +62,24 @@ public class TestClass {
     }
 
     @Then("^I see text \"([^\"]*)\"$")
-    public void iSeeText(String arg0) throws Throwable {
+    public void iSeeText(String text) {
         try {
-            driver.findElement(By.cssSelector("h2.a-spacing-none")).isDisplayed();
+            driver.findElement(By.cssSelector("h2.a-spacing-none")).toString().equals(text);
         } catch (NoSuchElementException e) {
             System.out.println("Login Unsuccessful");
             System.exit(1);
         }
     }
 
-    @When("^I fill \"([^\"]*)\" as \"([^\"]*)\"$")
-    public void itemSearch(String str1, String str2) {
+    @When("^I fill search as \"([^\"]*)\"$")
+    public void itemSearch(String searchWord) {
 
         driver.findElement(By.id("twotabsearchtextbox")).clear();
-        driver.findElement(By.id("twotabsearchtextbox")).sendKeys("Samsung");
+        driver.findElement(By.id("twotabsearchtextbox")).sendKeys(searchWord);
         driver.findElement(By.cssSelector(".nav-search-submit > input:nth-child(2)")).click();
         driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
 
-        int count = driver.findElements(By.linkText("Samsung")).size();
+        int count = driver.findElements(By.linkText(searchWord)).size();
 
         try {
             Assert.assertTrue(count >= 1);
@@ -89,9 +88,14 @@ public class TestClass {
         }
     }
 
+    /*
+    Element could not be found by partial link text. Hence, if else statement is used to identify the input.
+     */
     @When("^I navigate to \"([^\"]*)\"$")
-    public void goToPageTwo(String str3) {
-        driver.findElement(By.cssSelector("span.pagnLink:nth-child(3) > a:nth-child(1)")).click();
+    public void goToPageTwo(String pageNo) {
+        if ("Page 2".equals(pageNo)) {
+            driver.findElement(By.cssSelector("span.pagnLink:nth-child(3) > a:nth-child(1)")).click();
+        }
         driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
 
         try {
@@ -103,39 +107,39 @@ public class TestClass {
     }
 
     @When("^I select ([^\"]*) for ([^\"]*)$")
-    public void manageItem(String str4, String str5) {
-        driver.findElement(By.linkText("Click to see price")).click();
+    public void manageItem(String item, String list) {
+        if (item.isEmpty()) {
+            throw new NullPointerException();
+        } else{
+            driver.findElement(By.linkText("Click to see price")).click();
+        }
         driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
 
-        String urlCheck = new String(driver.findElement(By.className("a-size-large")).getText().toString());
-
-        driver.findElement(By.id("add-to-wishlist-button-submit")).click();
-        driver.manage().timeouts().implicitlyWait(2, TimeUnit.SECONDS);
-        driver.findElement(By.cssSelector("#WLHUC_viewlist > span:nth-child(1) > span:nth-child(1)")).click();
-        driver.manage().timeouts().implicitlyWait(5, TimeUnit.SECONDS);
-
-        try {
-            driver.findElement(By.linkText(urlCheck)).isDisplayed();
-        } catch (NoSuchElementException e) {
+        if ("WishList".equals(list)) {
+            driver.findElement(By.id("add-to-wishlist-button-submit")).click();
+            driver.manage().timeouts().implicitlyWait(2, TimeUnit.SECONDS);
+            driver.findElement(By.cssSelector("#WLHUC_viewlist > span:nth-child(1) > span:nth-child(1)")).click();
+            driver.manage().timeouts().implicitlyWait(5, TimeUnit.SECONDS);
+        }else {
             System.out.println("Element not found in Wish List");
-            System.exit(1);
+            throw new NullPointerException();
         }
     }
 
     @When("^I click (.+)$")
-    public void deleteItem(String str6){
-        driver.findElement(By.name("submit.deleteItem")).click();
-        try {
+    public void deleteItem(String button) {
+        if ("delete".equals(button)) {
+            driver.findElement(By.name("submit.deleteItem")).click();
+            driver.manage().timeouts().implicitlyWait(1,TimeUnit.SECONDS);
             driver.findElement(By.cssSelector("div.a-alert-inline:nth-child(1) > div:nth-child(1) > div:nth-child(2)")).isDisplayed();
-        } catch (NoSuchElementException e) {
+        }else {
             System.out.println("Element cannot be removed");
-            System.exit(1);
+            throw new NoSuchElementException();
         }
-
     }
 
-    @After
-    public void tearDown() {
+    @Then("^I quit (.+)$")
+    public void tearDown(String browser) {
         driver.manage().timeouts().implicitlyWait(5, TimeUnit.SECONDS);
         driver.quit();
         String verificationErrorString = verificationErrors.toString();
